Modernize ocular setup
